Add database migration

+++ spk/wallabag/src/installer.sh
@@ -122,6 +122,13 @@ postupgrade ()
 {
     mv ${TMP_DIR}/${PACKAGE}/parameters.yml ${CFG_FILE}
     mv -f ${TMP_DIR}/${PACKAGE}/db ${WEB_DIR}/${PACKAGE}/data/
+
+    # migrate database
+    if ! ${PHP} ${WEB_DIR}/${PACKAGE}/bin/console doctrine:migrations:migrate --env=prod -n -vvv > ${WEB_DIR}/${PACKAGE}/migration.log 2>&1; then
+        echo "Unable to migrate database schema. Please check the log: ${WEB_DIR}/${PACKAGE}/migration.log"
+        exit 1
+    fi
+
     rm -rf ${TMP_DIR}/${PACKAGE}
     exit 0
 }
